removed searchbar component and test page and combined into information page
