+++ src/App.js
@@ -2,7 +2,6 @@ import { BrowserRouter, Route, Switch } from "react-router-dom";
 import { InfoProvider } from "./utils/InfoContext";
 
 import { Home } from "./pages/Home";
-import { Test } from "./pages/Test";
 import { Information } from "./pages/Information";
 
 import './App.css';
@@ -14,7 +13,6 @@ function App() {
         <InfoProvider>  
           <Switch>
             <Route path="/information" component={Information}/>
-            <Route path="/test" component={Test}/>
             <Route path="/" component={Home}/>
           </Switch>
         </InfoProvider>
