+++ src/pages/Home.jsx
@@ -1,9 +1,17 @@
-import React from 'react'
+import { React, useContext } from 'react'
+import { Link } from 'react-router-dom'
+import { InfoContext } from "../utils/InfoContext";
 
 export const Home = () => {
+    const { info, setInfo } = useContext(InfoContext);
+
     return (
         <div>
-            This is home
+            <h1> Home Page </h1>
+            <Link to="/information">Click here to go to inforamation page</Link> <br/>
+            <button onClick={() => setInfo(null)}>
+              Reset Cache
+            </button>
         </div>
     )
 }
